ThreadPool: Use std::mutex for m_headMutex and TicketSpinLock for m_tailMutex.
Fairness on contention of m_headMutex is of no concern because the jobs will be
processed anyway, no matter in which thread. Fairness on contention of
m_tailMutex is still of some concern. Although only at most one worker thread
will access it at the same time job schedulers can still cause contention. We
have now addressed this using a TicketSpinLock, because the waits should be
relatively small.