Datum: Fixed incorrect exceptions being thrown.
An ErrnoException was thrown with errno although there is no reason expect that
errno is set, since we are dealing with the C++ standard library iostreams, not
doing our own C or POSIX I/O.

Breakage was introduced by previous commit (8564b1141e). This commit reverts to
the previous behavior of dealing with standard iostreams exceptions instead.