Optional: Removed the value() method.
Users might expect value() to throw an exception if there is no value contained.
This is the behavior of std::optional<T>::value() in C++17, but the current
implementation didn't throw in such cases, neither do we want it to. Hence, to
avoid confusion about this and to avoid introducing any bugs, removing these
value() getters seems to be the most reasonable option, especially when
identical functionality is already provided by operator*().