StrongType: Added StrongType(Pre|Post)?(In|De)crementable mixins.
Note that it currently not possible to use either both
StrongTypePreIncrement and StrongTypePostIncrement or both
StrongTypePreDecrement and StrongTypePostDecrement together, because in the
current naive solution the prefix and postfix implementations will end up in
different superclasses, which will break name lookup for operators ++ and --
before the () vs (int) overload resolution has the chance to kick in. This seems
to be a C++ language design flaw related to the overloading of prefix/postfix ++
and -- operators.

See https://stackoverflow.com/q/54091513/3919155 for a discussion.

In principle, this can somewhat be remedied by trying to merge the respective
mixins together in StrongType itself, but it would be very verbose and complex,
and hence we omitted it. When both prefix and postfix variants are required, we
suggest to use StrongTypeIncrement and/or StrongTypeDecrement instead.