NetworkMessage: Write strings using regular size instead. Breaks API.
This means that the maximum size of strings is now bound by std::uint64_t
instead of std::uint16_t. This might not be necessary in practice but it
simplifies the code quite a bit. The network overhead caused by this only
affects components using strings.

Removed the StringSizeType member type alias and the StringLengthError member
type.