ThreadPool: Removed useless OneThreadSharedSlice::Internal destructor.
In the destructor of OneThreadSharedSlice, m_internal is known to hold a valid
pointer. Tasks currently executing in the OneThreadSharedSlice may also hold a
valid pointer via Internal::m_sliceTask, but that is no longer the case after
we call m_internal->stopAndJoin(), because its post-condition guarantees that
no other task is queued for execution and no other task will be queued for
execution, meaning that m_internal will be the only live std::shared_ptr
pointing to internal, and that Internal will only be destroyed in the
destructor of OneThreadSharedSlice. Therefore we do not need to call
stopAndJoin() again in the destructor of Internal.